+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetop.py
@@ -100,7 +100,7 @@ class CreateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-09-08T15:14:50.397Z")
+    updated: Optional[datetime] = parse_datetime("2026-09-08T23:53:22.409Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")