+++ packages/orq-rc/RELEASES.md
@@ -10759,3 +10759,13 @@ Based on:
 - [python v4.15.0-rc.29] packages/orq-rc
 ### Releases
 - [PyPI v4.15.0-rc.29] https://pypi.org/project/orq-ai-sdk/4.15.0-rc.29 - packages/orq-rc
+
+## 2026-09-08 23:56:06
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.796.4 (2.935.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.15.0-rc.30] packages/orq-rc
+### Releases
+- [PyPI v4.15.0-rc.30] https://pypi.org/project/orq-ai-sdk/4.15.0-rc.30 - packages/orq-rc